+++ src/App.tsx
@@ -6,11 +6,14 @@ import { InstallPanel } from './components/InstallPanel'
 import { PinnedItemsPanel } from './components/PinnedItemsPanel'
 import { ResultsPanel } from './components/ResultsPanel'
 import { SearchPanel } from './components/SearchPanel'
+import { UpdatePanel } from './components/UpdatePanel'
 import { useAppViewModel } from './hooks/useAppViewModel'
+import { useUpdateCheck } from './hooks/useUpdateCheck'
 
 export default function App() {
   const { detailPanel, featuredSetsPanel, header, pinnedPanel, resultsPanel, searchPanel } =
     useAppViewModel()
+  const updateCheck = useUpdateCheck()
 
   return (
     <div className="app-shell">
@@ -20,6 +23,16 @@ export default function App() {
       />
 
       <InstallPanel />
+
+      <UpdatePanel
+        currentBuildId={updateCheck.currentBuildId}
+        currentVersion={updateCheck.currentVersion}
+        isUpdateAvailable={updateCheck.isUpdateAvailable}
+        latestBuildId={updateCheck.latestBuildId}
+        latestVersion={updateCheck.latestVersion}
+        onDismiss={updateCheck.acknowledgeUpdate}
+        onReload={updateCheck.reloadToUpdate}
+      />
 
       <SearchPanel {...searchPanel} />