+++ src/constants/app.ts
@@ -1,1 +1,2 @@
+export const APP_BUILD_ID = __APP_BUILD_ID__
 export const APP_VERSION = __APP_VERSION__